+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Delegs.hs
@@ -99,13 +99,13 @@ import NoThunks.Class (NoThunks (..))
 import Validation (failureUnless)
 
 data DelegsEnv era = DelegsEnv
-  { delegsSlotNo :: !SlotNo
+  { delegsSlotNo :: SlotNo
   , delegsEpochNo :: EpochNo
   -- ^ Lazy on purpose, because not all certificates need to know the current EpochNo
-  , delegsIx :: !TxIx
-  , delegspp :: !(PParams era)
-  , delegsTx :: !(Tx era)
-  , delegsAccount :: !AccountState
+  , delegsIx :: TxIx
+  , delegspp :: PParams era
+  , delegsTx :: Tx era
+  , delegsAccount :: AccountState
   }
 
 deriving stock instance
@@ -117,12 +117,12 @@ deriving stock instance
 data ShelleyDelegsPredFailure era
   = -- | Target pool which is not registered
     DelegateeNotRegisteredDELEG
-      !(KeyHash 'StakePool)
+      (KeyHash 'StakePool)
   | -- | Withdrawals that are missing or do not withdrawal the entire amount
     WithdrawalsNotInRewardsDELEGS
-      !(Map RewardAccount Coin)
+      (Map RewardAccount Coin)
   | -- | Subtransition Failures
-    DelplFailure !(PredicateFailure (EraRule "DELPL" era))
+    DelplFailure (PredicateFailure (EraRule "DELPL" era))
   deriving (Generic)
 
 type instance EraRuleFailure "DELEGS" ShelleyEra = ShelleyDelegsPredFailure ShelleyEra