+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Snap.hs
@@ -69,7 +69,7 @@ deriving instance Eq (SnapEvent era)
 
 instance NFData (SnapEvent era)
 
-data SnapEnv era = SnapEnv !(LedgerState era) !(PParams era)
+data SnapEnv era = SnapEnv (LedgerState era) (PParams era)
 
 instance EraTxOut era => STS (ShelleySNAP era) where
   type State (ShelleySNAP era) = SnapShots