+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/GenState.hs
@@ -212,8 +212,8 @@ data GenSize = GenSize
   deriving (Show)
 
 data GenEnv era = GenEnv
-  { gePParams :: !(PParams era)
-  , geSize :: !GenSize
+  { gePParams :: PParams era
+  , geSize :: GenSize
   }
 
 data GenState era = GenState